Updated the TwitterDAO integration tests and finished making the unit test to test showTweet

+++ core_java/twitter/src/test/java/ca/jrvs/apps/twitter/dao/TwitterDaoIntTest.java
@@ -66,8 +66,12 @@ public class TwitterDaoIntTest {
 
     @AfterClass
     public static void tearDown() {
-        twitterDao.deleteById(createTweetId.toString());
-        twitterDao.deleteById(showTweetId.toString());
+        if(createTweetId != null) {
+            twitterDao.deleteById(createTweetId.toString());
+        }
+        if(showTweetId != null) {
+            twitterDao.deleteById(showTweetId.toString());
+        }
     }
 
     @Test
